Create ButtonIcon, improve docs
